+++ GraphicsDevelopmentCoursework/Painter.hpp
@@ -51,11 +51,11 @@ class Painter {
 				bool addColor(std::string name, std::vector<float> color);
 			// creating vertexattribarrays
 				void useColor(std::string name);
-				void useVertices(unsigned int index);
+				void useVertices(size_t index);
 				void useColor(std::vector<float> colors);
 				void useVertices(std::vector<float> vertices);
 			// draw vertices
-				void drawVertices(unsigned int index); // the parameter is not used for we actually draw only one element
+				void drawVertices(size_t index); // the parameter is not used for we actually draw only one element
 			// cleaning
 				void disableVertexAttribArrays();
 };
